requirements-dev.txt
# Development dependencies
# Install with: pip install -r requirements-dev.txt

# Code quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Development tools
ipython>=8.15.0
django-debug-toolbar>=4.2.0

